Rename DanswerBot (#644)
