+++ backend/danswer/danswerbot/slack/blocks.py
@@ -8,15 +8,15 @@ from slack_sdk.models.blocks import DividerBlock
 from slack_sdk.models.blocks import HeaderBlock
 from slack_sdk.models.blocks import SectionBlock
 
-from danswer.bots.slack.constants import DISLIKE_BLOCK_ACTION_ID
-from danswer.bots.slack.constants import LIKE_BLOCK_ACTION_ID
-from danswer.bots.slack.utils import build_feedback_block_id
-from danswer.bots.slack.utils import remove_slack_text_interactions
-from danswer.bots.slack.utils import translate_vespa_highlight_to_slack
 from danswer.configs.constants import DocumentSource
 from danswer.configs.constants import SearchFeedbackType
 from danswer.configs.danswerbot_configs import DANSWER_BOT_NUM_DOCS_TO_DISPLAY
 from danswer.configs.danswerbot_configs import ENABLE_SLACK_DOC_FEEDBACK
+from danswer.danswerbot.slack.constants import DISLIKE_BLOCK_ACTION_ID
+from danswer.danswerbot.slack.constants import LIKE_BLOCK_ACTION_ID
+from danswer.danswerbot.slack.utils import build_feedback_block_id
+from danswer.danswerbot.slack.utils import remove_slack_text_interactions
+from danswer.danswerbot.slack.utils import translate_vespa_highlight_to_slack
 from danswer.direct_qa.interfaces import DanswerQuote
 from danswer.server.models import SearchDoc
 from danswer.utils.text_processing import replace_whitespaces_w_space
