+++ backend/danswer/danswerbot/slack/listener.py
@@ -9,18 +9,18 @@ from slack_sdk.socket_mode.request import SocketModeRequest
 from slack_sdk.socket_mode.response import SocketModeResponse
 from sqlalchemy.orm import Session
 
-from danswer.bots.slack.config import get_slack_bot_config_for_channel
-from danswer.bots.slack.constants import SLACK_CHANNEL_ID
-from danswer.bots.slack.handlers.handle_feedback import handle_slack_feedback
-from danswer.bots.slack.handlers.handle_message import handle_message
-from danswer.bots.slack.models import SlackMessageInfo
-from danswer.bots.slack.tokens import fetch_tokens
-from danswer.bots.slack.utils import ChannelIdAdapter
-from danswer.bots.slack.utils import decompose_block_id
-from danswer.bots.slack.utils import get_channel_name_from_id
-from danswer.bots.slack.utils import respond_in_thread
 from danswer.configs.danswerbot_configs import DANSWER_BOT_RESPOND_EVERY_CHANNEL
 from danswer.configs.danswerbot_configs import NOTIFY_SLACKBOT_NO_ANSWER
+from danswer.danswerbot.slack.config import get_slack_bot_config_for_channel
+from danswer.danswerbot.slack.constants import SLACK_CHANNEL_ID
+from danswer.danswerbot.slack.handlers.handle_feedback import handle_slack_feedback
+from danswer.danswerbot.slack.handlers.handle_message import handle_message
+from danswer.danswerbot.slack.models import SlackMessageInfo
+from danswer.danswerbot.slack.tokens import fetch_tokens
+from danswer.danswerbot.slack.utils import ChannelIdAdapter
+from danswer.danswerbot.slack.utils import decompose_block_id
+from danswer.danswerbot.slack.utils import get_channel_name_from_id
+from danswer.danswerbot.slack.utils import respond_in_thread
 from danswer.db.engine import get_sqlalchemy_engine
 from danswer.dynamic_configs.interface import ConfigNotFoundError
 from danswer.utils.logger import setup_logger
