+++ backend/danswer/danswerbot/slack/utils.py
@@ -12,12 +12,12 @@ from slack_sdk.errors import SlackApiError
 from slack_sdk.models.blocks import Block
 from slack_sdk.models.metadata import Metadata
 
-from danswer.bots.slack.constants import SLACK_CHANNEL_ID
-from danswer.bots.slack.tokens import fetch_tokens
 from danswer.configs.constants import ID_SEPARATOR
 from danswer.configs.danswerbot_configs import DANSWER_BOT_NUM_RETRIES
 from danswer.connectors.slack.utils import make_slack_api_rate_limited
 from danswer.connectors.slack.utils import SlackTextCleaner
+from danswer.danswerbot.slack.constants import SLACK_CHANNEL_ID
+from danswer.danswerbot.slack.tokens import fetch_tokens
 from danswer.utils.logger import setup_logger
 from danswer.utils.text_processing import replace_whitespaces_w_space
