+++ backend/supervisord.conf
@@ -22,7 +22,7 @@ autorestart=true
 # relatively compute-light (e.g. they tend to just make a bunch of requests to 
 # Vespa / Postgres)
 [program:celery_worker]
-command=celery -A danswer.background.celery worker --pool=threads --loglevel=INFO --logfile=/var/log/celery_worker.log
+command=celery -A danswer.background.celery worker --pool=threads --autoscale=3,10 --loglevel=INFO --logfile=/var/log/celery_worker.log
 stdout_logfile=/var/log/celery_worker_supervisor.log
 stdout_logfile_maxbytes=52428800
 redirect_stderr=true
@@ -41,7 +41,7 @@ autorestart=true
 # If not setup, this will just fail 5 times and then stop.
 # More details on setup here: https://docs.danswer.dev/slack_bot_setup
 [program:slack_bot_listener]
-command=python danswer/bots/slack/listener.py
+command=python danswer/danswerbot/slack/listener.py
 stdout_logfile=/var/log/slack_bot_listener.log
 stdout_logfile_maxbytes=52428800
 redirect_stderr=true
